+++ src/control_loop.cpp
@@ -192,24 +192,24 @@ void ControlLoop<JointPositions>::convertMotion(
     const RobotState& robot_state,
     research_interface::robot::MotionGeneratorCommand& command) {
   command.q_c = motion.q;
-  std::array<double, 7> previous_joint_position{};
+  std::array<double, 7> reference_position{};
   if (!initialized_filter_) {
-    previous_joint_position = command.q_c;
+    reference_position = command.q_c;
     initialized_filter_ = true;
   } else {
-    previous_joint_position = robot_state.q_d;
+    reference_position = robot_state.q_d;
   }
   if (cutoff_frequency_ < kMaxCutoffFrequency) {
     for (size_t i = 0; i < 7; i++) {
       command.q_c[i] =
-          lowpassFilter(kDeltaT, command.q_c[i], previous_joint_position[i], cutoff_frequency_);
+          lowpassFilter(kDeltaT, command.q_c[i], reference_position[i], cutoff_frequency_);
     }
   }
   if (limit_rate_) {
-    command.q_c =
-        limitRate(computeUpperLimitsJointVelocity(robot_state.q_d),
-                  computeLowerLimitsJointVelocity(robot_state.q_d), kMaxJointAcceleration,
-                  kMaxJointJerk, command.q_c, robot_state.q_d, robot_state.dq_d, robot_state.ddq_d);
+    command.q_c = limitRate(robot_.getUpperJointVelocityLimits(reference_position),
+                            robot_.getLowerJointVelocityLimits(reference_position),
+                            kMaxJointAcceleration, kMaxJointJerk, command.q_c, reference_position,
+                            robot_state.dq_d, robot_state.ddq_d);
   }
   checkFinite(command.q_c);
 }
@@ -228,8 +228,8 @@ void ControlLoop<JointVelocities>::convertMotion(
   }
   if (limit_rate_) {
     command.dq_c =
-        limitRate(computeUpperLimitsJointVelocity(robot_state.q_d),
-                  computeLowerLimitsJointVelocity(robot_state.q_d), kMaxJointAcceleration,
+        limitRate(robot_.getUpperJointVelocityLimits(robot_state.q_d),
+                  robot_.getLowerJointVelocityLimits(robot_state.q_d), kMaxJointAcceleration,
                   kMaxJointJerk, command.dq_c, robot_state.dq_d, robot_state.ddq_d);
   }
   checkFinite(command.dq_c);
@@ -241,30 +241,30 @@ void ControlLoop<CartesianPose>::convertMotion(
     const RobotState& robot_state,
     research_interface::robot::MotionGeneratorCommand& command) {
   command.O_T_EE_c = motion.O_T_EE;
-  std::array<double, 16> previous_cartesian_pose{};
-  std::array<double, 2> previous_elbow_pose{};
+  std::array<double, 16> reference_cartesian_pose{};
+  std::array<double, 2> reference_elbow_pose{};
   if (!initialized_filter_) {
-    previous_cartesian_pose = command.O_T_EE_c;
+    reference_cartesian_pose = command.O_T_EE_c;
     if (motion.hasElbow()) {
-      previous_elbow_pose = motion.elbow;
+      reference_elbow_pose = motion.elbow;
     }
     initialized_filter_ = true;
   } else {
-    previous_cartesian_pose = robot_state.O_T_EE_c;
+    reference_cartesian_pose = robot_state.O_T_EE_c;
     if (motion.hasElbow()) {
-      previous_elbow_pose = robot_state.elbow_c;
+      reference_elbow_pose = robot_state.elbow_c;
     }
   }
 
   if (cutoff_frequency_ < kMaxCutoffFrequency) {
-    command.O_T_EE_c = cartesianLowpassFilter(kDeltaT, command.O_T_EE_c, previous_cartesian_pose,
+    command.O_T_EE_c = cartesianLowpassFilter(kDeltaT, command.O_T_EE_c, reference_cartesian_pose,
                                               cutoff_frequency_);
   }
   if (limit_rate_) {
     command.O_T_EE_c = limitRate(
         kMaxTranslationalVelocity, kMaxTranslationalAcceleration, kMaxTranslationalJerk,
         kMaxRotationalVelocity, kMaxRotationalAcceleration, kMaxRotationalJerk, command.O_T_EE_c,
-        robot_state.O_T_EE_c, robot_state.O_dP_EE_c, robot_state.O_ddP_EE_c);
+        reference_cartesian_pose, robot_state.O_dP_EE_c, robot_state.O_ddP_EE_c);
   }
   checkMatrix(command.O_T_EE_c);
 
@@ -273,11 +273,11 @@ void ControlLoop<CartesianPose>::convertMotion(
     command.elbow_c = motion.elbow;
     if (cutoff_frequency_ < kMaxCutoffFrequency) {
       command.elbow_c[0] =
-          lowpassFilter(kDeltaT, command.elbow_c[0], previous_elbow_pose[0], cutoff_frequency_);
+          lowpassFilter(kDeltaT, command.elbow_c[0], reference_elbow_pose[0], cutoff_frequency_);
     }
     if (limit_rate_) {
       command.elbow_c[0] = limitRate(kMaxElbowVelocity, kMinElbowVelocity, kMaxElbowAcceleration,
-                                     kMaxElbowJerk, command.elbow_c[0], robot_state.elbow_c[0],
+                                     kMaxElbowJerk, command.elbow_c[0], reference_elbow_pose[0],
                                      robot_state.delbow_c[0], robot_state.ddelbow_c[0]);
     }
     checkElbow(command.elbow_c);