+++ test/mock_server.h
@@ -104,6 +104,8 @@ class MockServer {
  private:
   void serverThread();
   void sendInitialState(Socket& udp_socket);
+  void initializeServer();
+  void addDefaultRobotModelHandler();  // Specialized for RobotTypes
 
   template <typename T>
   MockServer& onSendUDP(std::function<T()> on_send_udp);