+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix: Fixed the elbow limit rating taking the wrong values
 - Breaking: Added the 'async position motion generator' control scheme, which allows for asynchronous position control of the robot
 - Feat: expose accelerometer data in robot state
+- Feat: Parse joint position–based velocity limits from URDF to compute upper and lower velocity bounds
 
 ## 0.16.0 - 2025-09-15